backend/requirements.txt
fastapi
uvicorn
python-multipart
sentence-transformers
faiss-cpu
langchain
langchain-community
greenlet
numpy
